+++ src/analysis/Interpreter.jl
@@ -24,18 +24,21 @@ struct LSInterpreter{S<:Server} <: JET.ConcreteInterpreter
     analyzer::LSAnalyzer
     counter::Counter
     activation_done::Union{Nothing,Base.Event}
+    warning_reports::Vector{JETLS.ToplevelWarningReport}
     state::JET.InterpretationState
     function LSInterpreter(
             server::S, request::AnalysisRequest, analyzer::LSAnalyzer, counter::Counter,
             activation_done::Union{Nothing,Base.Event}
         ) where S<:Server
-        return new{S}(server, request, analyzer, counter, activation_done)
+        return new{S}(server, request, analyzer, counter, activation_done, JETLS.ToplevelWarningReport[])
     end
     function LSInterpreter(
             server::S, request::AnalysisRequest, analyzer::LSAnalyzer, counter::Counter,
-            activation_done::Union{Nothing,Base.Event}, state::JET.InterpretationState
+            activation_done::Union{Nothing,Base.Event},
+            warning_reports::Vector{JETLS.ToplevelWarningReport},
+            state::JET.InterpretationState,
         ) where S<:Server
-        return new{S}(server, request, analyzer, counter, activation_done, state)
+        return new{S}(server, request, analyzer, counter, activation_done, warning_reports, state)
     end
 end
 
@@ -52,7 +55,7 @@ JET.InterpretationState(interp::LSInterpreter) = interp.state
 function JET.ConcreteInterpreter(interp::LSInterpreter, state::JET.InterpretationState)
     return LSInterpreter(
         interp.server, interp.request, interp.analyzer, interp.counter,
-        interp.activation_done, state)
+        interp.activation_done, interp.warning_reports, state)
 end
 JET.ToplevelAbstractAnalyzer(interp::LSInterpreter) = interp.analyzer
 function JET.ToplevelAbstractAnalyzer(
@@ -70,6 +73,27 @@ end
 
 # overloads
 # =========
+
+struct MethodDefinitionInfo
+    filename::String
+    mod::Module
+    src::Core.CodeInfo
+end
+
+const empty_lines_range = typemax(Int) => typemin(Int)
+
+function get_lines_in_src(filepath::AbstractString, src::Core.CodeInfo)
+    lines = empty_lines_range
+    for pc in 1:length(src.code)
+        lins = Base.IRShow.buildLineInfoNode(src.debuginfo, nothing, pc)
+        for lin in lins
+            if filepath == String(lin.file)
+                lines = min(lin.line,first(lines)) => max(lin.line,last(lines))
+            end
+        end
+    end
+    return lines
+end
 
 mutable struct SignatureAnalysisProgress
     const reports::Vector{JET.InferenceErrorReport}
@@ -89,7 +113,7 @@ end
 
 function cache_intermediate_analysis_result!(interp::LSInterpreter)
     result = JET.JETToplevelResult(interp.analyzer, interp.state.res, "LSInterpreter (intermediate result)", ())
-    intermediate_result = JETLS.new_analysis_result(interp.request, result)
+    intermediate_result = JETLS.new_analysis_result(interp, interp.request, result)
     JETLS.update_analysis_cache!(interp.server.state.analysis_manager, intermediate_result)
 end
 
@@ -108,8 +132,23 @@ function JET.analyze_from_definitions!(interp::LSInterpreter, config::JET.Toplev
     res = JET.InterpretationState(interp).res
     reset_report_target_modules!(interp.analyzer, res.analyzed_files)
 
+    # Detect method overwrites
+    seen_sigs = IdDict{Type,MethodDefinitionInfo}()
+    for signature_info in res.signature_infos
+        (; filename, mod, tt, src) = signature_info
+        if haskey(seen_sigs, tt)
+            lines = get_lines_in_src(filename, src)
+            original_definition = seen_sigs[tt]
+            original_filename = original_definition.filename
+            original_lines = get_lines_in_src(original_filename, original_definition.src)
+            push!(interp.warning_reports, JETLS.MethodOverwriteReport(mod, tt, filename, lines, original_filename, original_lines))
+        else
+            seen_sigs[tt] = MethodDefinitionInfo(filename, mod, src)
+        end
+    end
+
     entrypoint = config.analyze_from_definitions
-    n_sigs = length(res.toplevel_signatures)
+    n_sigs = length(res.signature_infos)
     n_sigs == 0 && return
     cancellable_token = interp.request.cancellable_token
     if cancellable_token !== nothing
@@ -131,7 +170,7 @@ function JET.analyze_from_definitions!(interp::LSInterpreter, config::JET.Toplev
             if cancellable_token !== nothing && is_cancelled(cancellable_token.cancel_flag)
                 return
             end
-            tt = res.toplevel_signatures[i]
+            (; tt) = res.signature_infos[i]
             # Create a new analyzer with fresh local caches (`inf_cache` and `analysis_results`)
             # to avoid data races between concurrent signature analysis tasks
             analyzer = JET.ToplevelAbstractAnalyzer(interp, JET.non_toplevel_concretized;